back/requirements.txt
Django==4.2.13
dj-database-url==1.2.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-markdownx==3.0.1
mysqlclient
python-dotenv
requests